+++ src/connection/mod.rs
@@ -1,15 +1,11 @@
 mod base;
 mod manager;
 
-pub use base::Connection;
-pub use manager::Manager;
+pub mod location;
 
-cfg_if::cfg_if! {
-    if #[cfg(unix)] {
-        mod unix;
-        pub use unix::Socket;
-    } else if #[cfg(windows)] {
-        mod windows;
-        pub use windows::Socket;
-    }
-}
+use base::Connection;
+pub(crate) use manager::Manager;
+
+mod os;
+
+pub use os::Socket;