+++ src/lib.rs
@@ -10,6 +10,12 @@
 
 //! A Rust library that allows the developer to interact with the Discord Presence API with ease
 
+use std::sync::atomic::AtomicBool;
+
+pub use client::Client;
+pub use error::{DiscordError, Result};
+pub use models::Event;
+
 pub(crate) static READY: AtomicBool = AtomicBool::new(false);
 
 // Cannot remove this *macro_use*, would break derive inside of macros
@@ -23,17 +29,15 @@ extern crate log;
 mod macros;
 /// A client for the Discord Presence API
 pub mod client;
-mod connection;
+/// Connection handling for the Discord Presence API
+pub mod connection;
 /// Errors that can occur when interacting with the Discord Presence API
 pub mod error;
 /// Event handlers
 pub mod event_handler;
 /// Models for discord activity
 pub mod models;
-mod utils;
 
-use std::sync::atomic::AtomicBool;
-
-pub use client::Client;
-pub use error::{DiscordError, Result};
-pub use models::Event;
+pub(crate) fn nonce() -> String {
+    uuid::Uuid::new_v4().to_string()
+}